backend/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
Flask-Migrate>=3.1
SQLAlchemy>=1.4
psycopg2-binary>=2.9
python-dotenv>=0.19
scipy>=1.7
flask-cors>=3.0
